pyproject.toml
[tool.poetry]
name = "dobot-python-api"
version = "0.1.0"
description = ""
authors = ["Andreas Riegg <andreas.riegg@gmx.de>"]
readme = "README.md"
packages = [{include = "dobot_python_api"}]

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
